Fix SuperSecretSetting
