+++ src/main/java/cc/hyperium/mixins/entity/MixinEntityRenderer.java
@@ -29,6 +29,7 @@ import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.entity.item.EntityItemFrame;
 import net.minecraft.util.*;
+import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Overwrite;
 import org.spongepowered.asm.mixin.Shadow;
@@ -65,17 +66,31 @@ public abstract class MixinEntityRenderer {
     @Shadow
     private Entity pointedEntity;
     private HyperiumEntityRenderer hyperiumEntityRenderer = new HyperiumEntityRenderer((EntityRenderer) (Object) this);
-
     @Shadow
     protected abstract void loadShader(ResourceLocation resourceLocationIn);
 
     @Shadow
     public abstract void stopUseShader();
 
+    @Shadow @Final public static int shaderCount;
+
+    @Shadow private int shaderIndex;
+
+    @Shadow @Final private static ResourceLocation[] shaderResourceLocations;
+
     //endStartSection
     @Inject(method = "updateCameraAndRender", at = @At(value = "INVOKE", target = "Lnet/minecraft/profiler/Profiler;endStartSection(Ljava/lang/String;)V", shift = At.Shift.AFTER))
     private void updateCameraAndRender(float partialTicks, long nano, CallbackInfo ci) {
         hyperiumEntityRenderer.updateCameraAndRender();
+    }
+
+    @Inject(method = "activateNextShader", at = @At("INVOKE_ASSIGN"))
+    public void activateNextShader(CallbackInfo callbackInfo){
+        if (this.shaderIndex != shaderCount) {
+            hyperiumEntityRenderer.useShader();
+        } else {
+            hyperiumEntityRenderer.stopShader();
+        }
     }
 
     /**
