+++ src/main/java/cc/hyperium/mixinsimp/entity/HyperiumEntityRenderer.java
@@ -26,9 +26,12 @@ import org.lwjgl.opengl.Display;
 
 public class HyperiumEntityRenderer {
     private EntityRenderer parent;
+    private boolean isUsingShader = false;
+    public static HyperiumEntityRenderer INSTANCE;
 
     public HyperiumEntityRenderer(EntityRenderer parent) {
         this.parent = parent;
+        INSTANCE = this;
     }
 
     public void updateCameraAndRender() {
@@ -189,5 +192,16 @@ public class HyperiumEntityRenderer {
                 }
             }
         }
+    }
+    public void useShader(){
+        isUsingShader = true;
+    }
+
+    public boolean isUsingShader(){
+        return isUsingShader;
+    }
+
+    public void stopShader(){
+        isUsingShader = false;
     }
 }
